requirements.txt
discord.py==2.1.0
yfinance==0.2.3
